+++ src/main/kotlin/com/dsoftware/ghmanager/ui/settings/GhActionsSettingsService.kt
@@ -7,7 +7,7 @@ data class GithubActionsManagerSettings(
     var useCustomRepos: Boolean = true,
     var customRepos: MutableMap<String, RepoSettings> = mutableMapOf(),
     var jobListAboveLogs: Boolean = true,
-    var frequency: Long = 30,
+    var frequency: Int = 30,
 ) {
     data class RepoSettings(
         var included: Boolean = true,
